sql/catalog/seqexpr: move util/sequence
This package is about looking inside expressions to find usages of sequence
functions for the purpose of dependency tracking. It definitely does not
belong in `util`. It can't go in `schemaexpr` where it would naturally belong
because it depends on builtins and builtins depend, indirectly, on schemaexpr.

Release justification: non-production code change

Release note: None
