+++ pkg/BUILD.bazel
@@ -199,6 +199,7 @@ ALL_TESTS = [
     "//pkg/sql/catalog/resolver:resolver_test",
     "//pkg/sql/catalog/schemadesc:schemadesc_test",
     "//pkg/sql/catalog/schemaexpr:schemaexpr_test",
+    "//pkg/sql/catalog/seqexpr:seqexpr_test",
     "//pkg/sql/catalog/systemschema:systemschema_test",
     "//pkg/sql/catalog/tabledesc:tabledesc_test",
     "//pkg/sql/catalog/typedesc:typedesc_test",
@@ -379,7 +380,6 @@ ALL_TESTS = [
     "//pkg/util/ring:ring_test",
     "//pkg/util/sdnotify:sdnotify_test",
     "//pkg/util/search:search_test",
-    "//pkg/util/sequence:sequence_test",
     "//pkg/util/shuffle:shuffle_test",
     "//pkg/util/span:span_test",
     "//pkg/util/stop:stop_test",
