+++ pkg/sql/schemachanger/scbuild/BUILD.bazel
@@ -21,6 +21,7 @@ go_library(
         "//pkg/sql/catalog/descs",
         "//pkg/sql/catalog/resolver",
         "//pkg/sql/catalog/schemaexpr",
+        "//pkg/sql/catalog/seqexpr",
         "//pkg/sql/catalog/tabledesc",
         "//pkg/sql/catalog/typedesc",
         "//pkg/sql/parser",
@@ -33,7 +34,6 @@ go_library(
         "//pkg/sql/sqltelemetry",
         "//pkg/util/errorutil/unimplemented",
         "//pkg/util/protoutil",
-        "//pkg/util/sequence",
         "@com_github_cockroachdb_errors//:errors",
         "@com_github_lib_pq//oid",
     ],
