+++ pkg/sql/schemachanger/scexec/scmutationexec/scmutationexec.go
@@ -16,13 +16,13 @@ import (
 
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/descpb"
+	"github.com/cockroachdb/cockroach/pkg/sql/catalog/seqexpr"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/tabledesc"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/typedesc"
 	"github.com/cockroachdb/cockroach/pkg/sql/parser"
 	"github.com/cockroachdb/cockroach/pkg/sql/schemachanger/scexec/descriptorutils"
 	"github.com/cockroachdb/cockroach/pkg/sql/schemachanger/scop"
 	"github.com/cockroachdb/cockroach/pkg/util/protoutil"
-	"github.com/cockroachdb/cockroach/pkg/util/sequence"
 	"github.com/cockroachdb/errors"
 )
 
@@ -113,7 +113,7 @@ func (m *visitor) UpdateRelationDeps(ctx context.Context, op scop.UpdateRelation
 			if err != nil {
 				return err
 			}
-			usedSequences, err := sequence.GetUsedSequences(expr)
+			usedSequences, err := seqexpr.GetUsedSequences(expr)
 			if err != nil {
 				return err
 			}
